move JS to .js files for proper linting
